HeaderValue.equalsIgnoringParameters & equalsOnlyPresentParameters
- promoted from HeaderValueWithParameters.
