+++ src/main/java/walkingkooka/net/header/HeaderValueWithParameters.java
@@ -34,15 +34,4 @@ public interface HeaderValueWithParameters<P extends HeaderParameterName<?>> ext
      * are different.
      */
     HeaderValueWithParameters<P> setParameters(final Map<P, Object> parameters);
-
-    /**
-     * Tests if two headers are equal ignoring any parameters on either.
-     */
-    boolean equalsIgnoringParameters(final Object other);
-
-    /**
-     * Tests if two headers are equal and requiring parameters on this to all be present on other, ignoring
-     * any extra parameters.
-     */
-    boolean equalsOnlyPresentParameters(final Object other);
 }
