+++ src/main/java/walkingkooka/net/header/HeaderValueWithParametersTesting.java
@@ -30,107 +30,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 public interface HeaderValueWithParametersTesting<V extends HeaderValueWithParameters<N>,
         N extends HeaderParameterName<?>> extends HeaderValueTesting<V> {
 
-    // equalsIgnoringParameters.........................................................................................
-
-    @Test
-    default void testEqualsIgnoringParametersNullFalse() {
-        equalsIgnoringParametersAndCheck(
-                this.createHeaderValueWithParameters(),
-                null,
-                false
-        );
-    }
-
-    @Test
-    default void testEqualsIgnoringParametersInvalidTypeFalse() {
-        equalsIgnoringParametersAndCheck(
-                this.createHeaderValueWithParameters(),
-                this,
-                false
-        );
-    }
-
-    @Test
-    default void testEqualsIgnoringParametersSelfTrue() {
-        final V header = this.createHeaderValueWithParameters();
-        equalsIgnoringParametersAndCheck(
-                header,
-                header,
-                true
-        );
-    }
-
-    @Test
-    default void testEqualsIgnoringParametersTrue() {
-        equalsIgnoringParametersAndCheck(
-                this.createHeaderValueWithParameters(),
-                this.createHeaderValueWithParameters(),
-                true
-        );
-    }
-
-    default void equalsIgnoringParametersAndCheck(final HeaderValueWithParameters<?> header,
-                                                  final Object other,
-                                                  final boolean expected) {
-        assertEquals(expected,
-                header.equalsIgnoringParameters(other),
-                () -> header + " equalsIgnoringParameters " + other);
-
-        if (other instanceof HeaderValueWithParameters) {
-            final HeaderValueWithParameters<?> otherHeader = Cast.to(other);
-            assertEquals(expected,
-                    otherHeader.equalsIgnoringParameters(header),
-                    () -> otherHeader + " equalsIgnoringParameters " + header);
-        }
-    }
-
-    // equalsOnlyPresentParameters.........................................................................................
-
-    @Test
-    default void testEqualsOnlyPresentParametersNullFalse() {
-        equalsOnlyPresentParametersAndCheck(
-                this.createHeaderValueWithParameters(),
-                null,
-                false
-        );
-    }
-
-    @Test
-    default void testEqualsOnlyPresentParametersInvalidTypeFalse() {
-        equalsOnlyPresentParametersAndCheck(
-                this.createHeaderValueWithParameters(),
-                this,
-                false
-        );
-    }
-
-    @Test
-    default void testEqualsOnlyPresentParametersSelfTrue() {
-        final V header = this.createHeaderValueWithParameters();
-        equalsOnlyPresentParametersAndCheck(
-                header,
-                header,
-                true
-        );
-    }
-
-    @Test
-    default void testEqualsOnlyPresentParametersTrue() {
-        equalsOnlyPresentParametersAndCheck(
-                this.createHeaderValueWithParameters(),
-                this.createHeaderValueWithParameters(),
-                true
-        );
-    }
-
-    default void equalsOnlyPresentParametersAndCheck(final HeaderValueWithParameters<?> header,
-                                                     final Object other,
-                                                     final boolean expected) {
-        assertEquals(expected,
-                header.equalsOnlyPresentParameters(other),
-                () -> header + " equalsOnlyPresentParameters " + other);
-    }
-
     // setParameters ...........................................................................................
 
     @Test
